+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased] - ReleaseDate
 
+## Added
+
+- Added the ability to turn off text node consolidation. Use
+  `Xot::set_text_consolidation(false)` to turn it off or on again.
+
 ## [0.11.0] - 2023-01-13
 
 ### Added
@@ -226,8 +231,8 @@
 Initial public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/faassen/xot/compare/v0.11.0...HEAD
 
+[unreleased]: https://github.com/faassen/xot/compare/v0.11.0...HEAD
 [unreleased]: https://github.com/faassen/xot/compare/v0.10.3...v0.11.0
 [unreleased]: https://github.com/faassen/xot/compare/v0.10.2...v0.10.3
 [unreleased]: https://github.com/faassen/xot/compare/v0.10.1...v0.10.2